Improve operators
Introduce a new null_safe_op() function which is used instead of
isbits() to decide whether to take the fast path (without a branch)
or the slow path in lifted operators. This is more correct (and thus
safer), and allows supporting operations on any type. Also add an
extensive set of tests for both paths.

On Julia 0.4, operators for which no functor exist cannot be identified
by null_safe_op(): for these, fall back to the slow path on that version.

This PR introduces cbrt as a new supported operation since it has a special
operator ∛. It also keeps abs(), abs2(), scalarmin() and scalarmax(), which
should probably be deprecated once we have a more general solution for
lifting.